Do not accept private keys longer than 32 bytes
If the user enters a key in hexadecimal format, it is possible that
they will accidentally type more than 32 bytes. If this happens, the
script will continue normally, but will produce invalid (and
unspendable) WIF keys and addresses.

Therefore, do not accept hexadecimal input longer than 32 bytes.
Shorter inputs are not a problem, since the private key will just
have 0's for the missing digits.